+++ java-security/src/main/java/com/sap/cloud/security/token/validation/validators/JsonWebKeyImpl.java
@@ -96,11 +96,7 @@ class JsonWebKeyImpl implements JsonWebKey {
 
 	@Override
 	public int hashCode() {
-		return calculateUniqueId(keyAlgorithm, keyId);
-	}
-
-	public static int calculateUniqueId(JwtSignatureAlgorithm algorithm, String keyId) {
-		return Objects.hash(algorithm, keyId != null ? keyId : DEFAULT_KEY_ID);
+		return Objects.hash(keyAlgorithm, keyId != null ? keyId : DEFAULT_KEY_ID);
 	}
 
 	@Override